Add support for PEP 503 "simple" API.
This adds support for the "simple" index API specified in PEP 503.
While PyPI now recommends using its JSON API, other indices do not yet
support that format and it has yet to be standardized.

The simple API provides less information than the JSON API, so some
refactoring was required to provide a common interface for both APIs to
the rest of the code. This interface is in package_indices.py. While I
was at it I grouped the other code for getting info on dists into a
single submodule.